+++ spec/support/stub_engine.rb
@@ -3,7 +3,12 @@
 module MissionControl
   module Jobs
     unless defined?(MissionControl::Jobs::Engine)
-      class Engine < Rails::Engine
+      class Engine < ::Rails::Engine
+        endpoint lambda { |_env|
+          [200,
+           { "content-type" => "text/html" },
+           ["<!DOCTYPE html><html><head><title>MC</title></head><body>Jobs</body></html>"]]
+        }
       end
     end
   end